Remove redundant logging. Declare merge variables as doubles not ints.

+++ planetiler-custommap/src/main/java/com/onthegomap/planetiler/custommap/configschema/MergeLineStrings.java
@@ -3,7 +3,7 @@ package com.onthegomap.planetiler.custommap.configschema;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record MergeLineStrings(
-  @JsonProperty("min_length") int minLength,
-  @JsonProperty() int tolerance,
-  @JsonProperty() int buffer
+  @JsonProperty("min_length") double minLength,
+  @JsonProperty() double tolerance,
+  @JsonProperty() double buffer
 ) {}
